requirements.txt
django
python-decouple
djangorestframework
django-cors-headers
drf-yasg
django-jazzmin==3.0.1
pillow
psycopg2-binary
drf-spectacular